requirements.txt
fastapi[all]==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.13.0
pytest==7.4.3
python-dotenv==1.0.0
